requirements.txt
configparser==5.0.2
requests==2.26.0
kafka-python==2.0.2
psycopg2-binary==2.9.3